\name{CheckParaTable}
\alias{CheckParaTable}
\title{
The method 'CheckParaTable' in \link{IPPModel} class
}

\usage{
CheckParaTable()
}

\description{
This method validates the information in \link{ParaTable}. For example, if the column 'distMeasure' is 'euclidean' or 'cosine', it is validated.
}
\value{
TRUE, if the information in \link{ParaTable} are validated; FALSE, otherwise.
}

\seealso{
\link{IPPModel}
}
