\name{Clustering.Res}
\alias{Clustering.Res}

\title{The field 'Clustering.Res' in \link{IPPModel} class}

\description{
Clustering.Res is the clustering results and is generate by method \link{ClusterImpactPlots}. Clustering.Res is a list, each element of which is a data.frame of the clustering results of a feature.
}

\seealso{
\link{IPPModel}
}
