\name{ColorList}
\alias{ColorList}

\title{The field 'ColorList' in \link{IPPModel} class}

\description{
ColorList is the list of curve colors used for drawing IPPs. Its default value is c("red", "darkgreen", "blue", "orange", "black", "purple", "pink", "cyan", "cadeblue"). If the default value is not changed and the number of curves in IPPs are greater than the number of colors, 'black' color is used for all of the redundant curves.
}

\seealso{
\link{IPPModel}
}
