\name{DrawIPP}
\alias{DrawIPP}

\title{
The method 'DrawIPP' in \link{IPPModel} class
}

\usage{
DrawIPP(centralized=TRUE, nc=4)
}

\arguments{
  \item{centralized}{a boolean, indicating if the impact pattern plots are centralized.}
  \item{nc}{an integer, the number of subplots that are shown in every row.}
 }

\description{
The method draws the impact pattern plots based on the clustering results \link{Clustering.Res}.
}

\seealso{
\link{IPPModel}
}

