\name{ExecuteAll}
\alias{ExecuteAll}

\title{
The method 'ExecuteAll' in \link{IPPModel} class
}

\usage{
ExecuteAll()
}

\description{
The method executes the methods \link{SamplingXA}, \link{SamplingXB}, \link{PredictData}, \link{ClusterImpactPlots} and \link{BuildTree} in sequence.
}

\seealso{
\link{IPPModel}
}

