\name{FIN.Data}
\alias{FIN.Data}

\title{The field 'FIN.Data' in \link{IPPModel} class}

\description{
FIN.Data is a data.frame to describe the feature interaction network, and is generated by the method \link{BuildTree}. The row names and column names denote the 'from' nodes and 'to' nodes of the network, respectively. The values in FIN.Data denotes the link weights.
}

\seealso{
\link{IPPModel}
}
