\name{ParaTable}
\alias{ParaTable}

\title{The field 'ParaTable' in \link{IPPModel} class}

\description{
The field 'ParaTable' provides some parameter information for the IPPModel class. It is a data.frame whose rowNames are the names of the input features in \link{X.Data}. It includs the following columns.
}

\section{Columns}{
\describe{
    \item{dataType}{string, the data types of input features. It should be "interval", "binary", "ordinal" or "nominal". Interval feature is one for which the mean (or average) makes sense, such as person's height. Binary feature has only two possible levels. Gender is an example. Nominal variable has more than two levels, but the values of the levels have no implied order. Colors are examples. Ordinal feature has more than two levels, and the values of the levels have an implied order. Coffee sizes, such as small, medium, and large, are examples.}
    \item{uniqueValue}{integer, the numbe of unique values of the input features.}
    \item{X_A}{boolean, indicating if the feature serves as a target feature to draw impact pattern plots (IPPs) and involves in the feature interaction network (FIN).}
    \item{L_A}{integer, the number of levels sampled from \link{X.Data} for the feature.}
    \item{samplingMethod}{string, the sampling method of the feature. It should be "equal", "percentile" or "random". "equal" means that all levels are sampled with equal width from the range of the feature. "percentile" means that all levels are sampled based on the percentiles of the feature's distribution. "random" means that all levels are sampled randomly without replacement.}
    \item{clusteringMethod}{string, the clustering method. It should be "kmeans" or "kmedoids".}
    \item{centralized}{boolean, indicating if the feature's impact plots are centralized before clustering.}
    \item{distMeasure}{string, the distance measure used in clustering process. It should be "euclidean" or "cosine".}
    \item{autoK}{boolean, indicating if the number of clusters is determined automatically based on the Dunn index.}
    \item{numK}{integer, the number of clusters if autoK = FALSE. If autoK = TRUE, numK denotes the maximum number of clusters.}
    \item{treeDepth}{integer, the maximum tree depth. It is used to control the size of decision tree.}
    \item{minSplit}{integer, the minimum number of observations for tree node splitting. It is used to control the size of decision tree.}
}
}

\section{Note}{
The ParaTable can be generated by the method \link{GenerateParaTable} based on the field \link{X.Data}. However, the user must check the parameter information carefully before continuing the other tasks.
}

\seealso{
\link{IPPModel}
}

