\name{Pred.Fun}
\alias{Pred.Fun}

\title{The field 'Pred.Fun' in \link{IPPModel} class}

\description{
Pred.fun is the prediction function. It can be any model created by "nnet", "randomforest" and "kernlab" etc. Pred.fun must be consistent with the field \link{X.Data}.
}

\seealso{
\link{IPPModel}
}
