\name{Pred.Res}
\alias{Pred.Res}

\title{The field 'Pred.Res' in \link{IPPModel} class}

\description{
Pred.Res is the prediction results of the model defined by \link{Pred.Fun} for the combination of \link{XA.Sample} and \link{XB.Sample}. Pred.Res is generated by the method \link{PredictData}. Pred.Res is a list, each element of which is a data.frame of the prediction results of a feature.
}

\seealso{
\link{IPPModel}
}
