\name{SamplingXA}
\alias{SamplingXA}

\title{
The method 'SamplingXA' in \link{IPPModel} class
}

\usage{
SamplingXA()
}

\description{
The method samples X_A from \link{X.Data} based on the sample size and the sampling method defined by "L_A" and "samplingMethod" in \link{ParaTable}, respectively.
}
\value{
a list of samples of each feature. The result is saved in the field \link{XA.Sample}.
}

\seealso{
\link{IPPModel}
}
