\name{SamplingXB}
\alias{SamplingXB}

\title{
The method 'SamplingXB' in \link{IPPModel} class
}

\usage{
SamplingXB()
}

\description{
The method samples X_B from \link{X.Data} based on the sample size and sampling method defined by the fields \link{XB.Size} and \link{XB.SamplingMethod}, respectively.
}
\value{
a data.frame of sample. It is saved in the field \link{XB.Sample}.
}

\seealso{
\link{IPPModel}
}
