\name{TaskFinishTime}
\alias{TaskFinishTime}

\title{The field 'TaskFinishTime' in \link{IPPModel} class}

\description{
TaskFinishTime is a list for recording the finishing time of tasks, including initialization, sampling X_A and X_B, prediction, clustering, obtaining tree rules and building FIN.
}

\seealso{
\link{IPPModel}
}
