\name{TreeRules}
\alias{TreeRules}

\title{The field 'TreeRules' in \link{IPPModel} class}

\description{
TreeRules are the decision tree rules, which is generated by method \link{BuildTree}. It is a list, each element of which is a data.frame of tree rules and the corresponding cluster distribution of a feature.
}

\seealso{
\link{IPPModel}
}
