\name{WriteToExcel}
\alias{WriteToExcel}

\title{
The method 'WriteToExcel' in \link{IPPModel} class
}

\usage{
WriteToExcel(excelName)
}

\arguments{
  \item{excelName}{a string, the path and name of the excel file.}
 }

\description{
The method writes some results into an excel file, including the impact pattern plots, the tree rules and the feature interaction network.
}

\seealso{
\link{IPPModel}
}

