\name{X.Data}
\alias{X.Data}

\title{The field 'X.Data' in \link{IPPModel} class}

\description{
    X.Data is a data.frame and is the dataset of input features. X.Data must be consistent with the field \link{Pred.Fun} that is the prediction function. Note that the dataset should not include the output feature (y).
}
\seealso{
\link{IPPModel}
}
