\name{XA.Sample}
\alias{XA.Sample}

\title{The field 'XA.Sample' in \link{IPPModel} class}

\description{
XA.Sample is the sample of X_A extracted from \link{X.Data}. It is generated by method \link{SamplingXA}. XA.Sample is a list, each element of which is a vector of the sample values of a feature. The included features in XA.Sample are determined by the column 'X_A' in the field \link{ParaTable}. The size of XA.Sample is defined by the column 'L_A' in the field \link{ParaTable}. The sampling method of XA.Sample is defined by the column 'samplingMethod' in the field \link{ParaTable}.
}

\seealso{
\link{IPPModel}
}
