\name{XB.Sample}
\alias{XB.Sample}

\title{The field 'XB.Sample' in \link{IPPModel} class}

\description{
XB.Sample is the sample of X_B extracted from \link{X.Data}. It is generated by method \link{SamplingXB}. The size of XB.Sample is defined by the field \link{XB.Size}. The sampling method of XB.Sample is defined by the field \link{XB.SamplingMethod}. XB.Sample is a data.frame whose columns are the same as \link{X.Data}.
}

\seealso{
\link{IPPModel}
}
