\name{XB.SamplingMethod}
\alias{XB.SamplingMethod}

\title{The field 'XB.SamplingMethod' in \link{IPPModel} class}

\description{
XB.SamplingMethod is the sampling method of \link{XB.Sample}, "joint" or "independent". "joint" means that all features are sampled from X.Data jointly. "independent" means that each feature is sampled independently; then all features are combined randomly. Note that the "joint" method can keep the features' joint distribution unchanged.
}

\seealso{
\link{IPPModel}
}
